ucs: Document and improve trimming algorithm

+++ Demo/exp430f5438/FreeRTOSConfig.h
@@ -76,7 +76,10 @@
 #define configUSE_PREEMPTION		1
 #define configUSE_IDLE_HOOK			1
 #define configUSE_TICK_HOOK			0
-#define configCPU_CLOCK_HZ			( ( unsigned long ) 8000000 ) /* Clock setup from main.c in the demo application. */
+//#define configCPU_CLOCK_HZ			1000000UL
+#define configCPU_CLOCK_HZ			8000000UL
+//#define configCPU_CLOCK_HZ			16000000UL
+//#define configCPU_CLOCK_HZ			20000000UL
 #define configTICK_RATE_HZ			( ( portTickType ) 1000 )
 #define configMAX_PRIORITIES		( ( unsigned portBASE_TYPE ) 4 )
 #define configMINIMAL_STACK_SIZE	( ( unsigned short ) 300 )
